Auto stash before merge of "master" and "raid_helper/master"
Added local database to provide more control over users and channels

+++ host_channel_creator/cogs/RaidHelper.py
@@ -1,6 +1,7 @@
 import discord
 import os
 from discord.ext import commands
+import sqlite3
 
 
 class RaidHelper(commands.Cog, discord.Client):
@@ -17,7 +18,44 @@ class RaidHelper(commands.Cog, discord.Client):
 
     @commands.Cog.listener()
     async def on_ready(self):
+        # Create the DB table if it does not exist on startup
+        db = sqlite3.connect('RaidHelper.sqlite').cursor()
+        db.execute(
+        '''
+        CREATE TABLE IF NOT EXISTS RaidHelper 
+        (
+            user_id	INTEGER NOT NULL,
+            user_name TEXT NOT NULL,
+            channel_id INTEGER NOT NULL,
+            channel_name INTEGER NOT NULL,
+            PRIMARY KEY(user_id)
+        )
+        ''')
+        db.close()
         print('RaidHelper cog is loaded.')
+
+    @commands.command()
+    @commands.has_role('Shiny Raid Host')
+    async def delete(self, ctx):
+        if ctx.message.author == self.bot_user:
+            return
+
+        await ctx.message.delete()
+        # Get user data from DB
+        db = sqlite3.connect('RaidHelper.sqlite')
+        cursor = db.cursor()
+        row = cursor.execute(f'SELECT * FROM RaidHelper WHERE user_id = {ctx.message.author.id}').fetchone()
+        # If there is any data, it means the user has a channel which can be deleted
+        if row:
+            await ctx.guild.get_channel(row[2]).delete()
+            await ctx.send("Channel " + row[3] + " has been deleted")
+            # Remove user data from DB
+            cursor.execute(f'DELETE FROM RaidHelper WHERE user_id = {row[0]}')
+            db.commit()
+        else:
+            await ctx.message.channel.send('You do not have any channel created.')
+        cursor.close()
+        db.close()
 
     @commands.command()
     @commands.has_role('Shiny Raid Host')
@@ -26,17 +64,21 @@ class RaidHelper(commands.Cog, discord.Client):
             return
         # If channel name is given
         if chan_name:
-            self.channel_creator = ctx.message.author
-            # Delete the user message and send an embed to the channel
-            await ctx.message.delete()
-            await ctx.message.channel.send(
-                embed=discord.Embed().set_author(name="Channel named '" + chan_name + "' has been created.",
-                                                 icon_url='https://cdn.discordapp.com/attachments/662128235982618635/704757893798428732/SeekPng.com_green-tick-icon-png_3672259.png'))
-            # Create a new channel based on category
-            category = discord.utils.get(ctx.guild.categories, name='Text Channels')
-            self.new_chan = await ctx.guild.create_text_channel(chan_name, category=category)
-            await self.new_chan.edit(overwrites=None)
-
+            db = sqlite3.connect('RaidHelper.sqlite')
+            cursor = db.cursor()
+            row = cursor.execute(f'SELECT user_id FROM RaidHelper WHERE user_id = {ctx.message.author.id}').fetchone()
+            # Empty row data means user does not already have a channel created
+            if row is None:
+                # Delete the user message and send an embed to the channel
+                await ctx.message.delete()
+                await ctx.message.channel.send(
+                    embed=discord.Embed().set_author(name="Channel named '" + chan_name + "' has been created.",
+                                                     icon_url='https://cdn.discordapp.com/attachments/662128235982618635/704757893798428732/SeekPng.com_green-tick-icon-png_3672259.png'))
+                # Create a new channel based on category
+                category = discord.utils.get(ctx.guild.categories, name='Text Channels')
+                new_chan = await ctx.guild.create_text_channel(chan_name, category=category)
+                # Set permissions for the user in this new channel
+                await new_chan.set_permissions(ctx.message.author, manage_messages=True)
             # Setup and command list
             commandembed = discord.Embed().set_author(name='List of Commands', icon_url='https://cdn.discordapp.com/attachments/704174855813070901/705216533923889263/491Darkrai.png')
             commandembed.add_field(name='$mute', value='Used to mute a player', inline=False)
@@ -44,9 +86,17 @@ class RaidHelper(commands.Cog, discord.Client):
             commandembed.add_field(name='$ban', value='Used to ban a player from the channel', inline=False)
             commandembed.add_field(name='$unban', value='Used to unban a player from the channel', inline=False)
             await self.new_chan.send(embed=commandembed)
-
-            # Set permissions for the user in this new channel
-            await self.new_chan.set_permissions(self.channel_creator, manage_messages=True)
+                # Write user and channel info into DB for later use
+                cursor.execute(
+                    """
+                    INSERT INTO RaidHelper (user_id, user_name, channel_id, channel_name) VALUES (?, ?, ?, ?)
+                    """,
+                    (ctx.message.author.id, ctx.message.author.display_name, new_chan.id, new_chan.name))
+                db.commit()
+            else:
+                await ctx.message.channel.send('You already have created channel created.')
+            cursor.close()
+            db.close()
         else:
             inputnameembed = discord.Embed(description='<:x_:705214517961031751> Please input a name for the channel after the command.')
             await ctx.message.channel.send(embed=inputnameembed)
